pyproject.toml
[project]
name = "hangman"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=44.0.3",
    "pillow>=11.2.1",
    "sqlalchemy>=2.0.40",
]
